+++ NEWS.md
@@ -10,7 +10,8 @@ of the display.
 edges of a mesh.
 * Added material property `tag`, a string associated 
 with each object.  The value is reported by `ids3d()` and
-may be used to select objects in `pop3d()`, but otherwise is 
+may be used to select objects in most functions that use ids,
+but otherwise is 
 largely ignored by `rgl`.  The `tagged3d()` function returns
 information on tags.